remove duplicate from TestBatch TestAppend and TestDelete
