+++ bigtable-hbase-parent/bigtable-hbase-integration-tests-common/src/main/java/com/google/cloud/bigtable/hbase/AbstractTestDelete.java
@@ -47,7 +47,7 @@ public abstract class AbstractTestDelete extends AbstractTest{
 
     table.close();
   }
-
+  
   @Test
   public void testDeleteEmptyRow() throws IOException {
     // Initialize data
